+++ controler/MailSecControler.class.php
@@ -9,12 +9,13 @@ class MailSecControler extends PastellControler
         parent::_beforeAction();
         $id_e = $this->getPostOrGetInfo()->getInt('id_e');
         $this->setViewParameter('id_e', $id_e);
-        $this->hasDroitLecture($id_e);
+        $this->hasEntiteDroitLecture($id_e);
         $this->setNavigationInfo($id_e, "MailSec/annuaire?");
         $this->setViewParameter('menu_gauche_select', 'MailSec/annuaire');
         $this->setViewParameter('menu_gauche_template', "EntiteMenuGauche");
         $this->setDroitLectureOnConnecteur($id_e);
         $this->setDroitImportExportConfig($id_e);
+        $this->setDroitLectureOnUtilisateur($id_e);
     }