+++ controler/PastellControler.class.php
@@ -33,6 +33,14 @@ class PastellControler extends Controler
         ));
     }
 
+    protected function setDroitLectureOnUtilisateur(int $id_e): void
+    {
+        $this->setViewParameter('droitLectureOnUtilisateur', $this->getDroitService()->hasDroitUtilisateurLecture(
+            $id_e,
+            $this->getId_u()
+        ));
+    }
+
     protected function setDroitImportExportConfig(int $id_e): void
     {
         $this->setViewParameter(
@@ -63,9 +71,18 @@ class PastellControler extends Controler
      * @throws LastMessageException
      * @throws LastErrorException
      */
-    public function hasDroitLecture($id_e)
+    public function hasUtilisateurDroitLecture(int $id_e): void
     {
-        $this->verifDroit($id_e, "entite:lecture");
+        $this->verifDroit($id_e, DroitService::getDroitLecture(DroitService::DROIT_UTILISATEUR));
+    }
+
+    /**
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    public function hasEntiteDroitLecture(int $id_e): void
+    {
+        $this->verifDroit($id_e, DroitService::getDroitLecture(DroitService::DROIT_ENTITE));
     }
 
     /**